requirements.txt
asgiref==3.5.0
astroid==2.9.3
colorama==0.4.4
dj-database-url==0.5.0
Django==4.0.2
djangorestframework==3.13.1
gunicorn==20.1.0
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
platformdirs==2.4.1
pylint==2.12.2
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.0.1
tzdata==2021.5
whitenoise==5.3.0
wrapt==1.13.3
